% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\docType{class}
\name{DeltaMergeBuilder-class}
\alias{DeltaMergeBuilder-class}
\title{S4 class that represents pending merge operation}
\description{
S4 class that represents pending merge operation
}
\section{Slots}{

\describe{
\item{\code{jmb}}{A Java object reference to the backing DeltaMergeBuilder}

\item{\code{.target}}{DeltaTable. A reference to the target table}
}}

\note{
DeltaMergeBuilder, since 1.0.0
}
\concept{DeltaMergeBuilder functions}
