% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{class}
\name{DeltaTable-class}
\alias{DeltaTable-class}
\title{S4 class that represents DeltaTable}
\description{
S4 class that represents DeltaTable
}
\section{Slots}{

\describe{
\item{\code{jdt}}{A Java object reference to the backing DeltaTable}
}}

\note{
DeltaTable since 1.0.0
}
\concept{DeltaTable functions}
