% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\docType{class}
\name{DeltaTableBuilder-class}
\alias{DeltaTableBuilder-class}
\title{S4 class that represents pending table build operation}
\description{
S4 class that represents pending table build operation
}
\section{Slots}{

\describe{
\item{\code{method}}{character name of the initializer method}

\item{\code{ops}}{list list of operations in form of list(.method = ..., args = list(...))}
}}

\note{
DeltaTableBuilder, since 1.0.0
}
\examples{
\dontrun{
  path <- tempfile()

  dlt_create() \%>\%
    dlt_location(path) \%>\%
    dlt_add_column("id", "integer", nullable = FALSE) \%>\%
    dlt_add_columns(structType("key string, value double")) \%>\%
    dlt_partitioned_by("key") \%>\%
    dlt_comment("Key-value table") \%>\%
    dlt_property("creation-time", as.character(Sys.time())) \%>\%
    dlt_execute()
}

}
\concept{DeltaTableBuilder functions}
