% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\name{dlt_add_columns,InternalDeltaTableBuilder,structType-method}
\alias{dlt_add_columns,InternalDeltaTableBuilder,structType-method}
\alias{dlt_add_columns,InternalDeltaTableBuilder,character-method}
\title{Add columns to the build table}
\usage{
\S4method{dlt_add_columns}{InternalDeltaTableBuilder,structType}(dtb, schema)

\S4method{dlt_add_columns}{InternalDeltaTableBuilder,character}(dtb, schema)
}
\arguments{
\item{dtb}{InternalDeltaTableBuilder}

\item{schema}{character (DDL string) or structType}
}
\value{
this InternalDeltaTableBuilder
}
\description{
Add columns to the build table
}
\section{Functions}{
\itemize{
\item \code{dlt_add_columns,InternalDeltaTableBuilder,structType-method}: Add columns to the table

\item \code{dlt_add_columns,InternalDeltaTableBuilder,character-method}: Add columns to the table
}}

\note{
dlt_add_columns, since 1.0.0
}
\keyword{internal}
