% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_add_columns,DeltaTableBuilder,ANY-method}
\alias{dlt_add_columns,DeltaTableBuilder,ANY-method}
\title{Add columns to the build table}
\usage{
\S4method{dlt_add_columns}{DeltaTableBuilder,ANY}(dtb, schema)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{schema}{character (DDL string) or structType}
}
\value{
DeltaTableBuilder
}
\description{
Add columns to the build table
}
\section{Functions}{
\itemize{
\item \code{dlt_add_columns,DeltaTableBuilder,ANY-method}: Add columns to the table
}}

\note{
dlt_add_columns, since 1.0.0
}
