% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\name{dlt_comment,InternalDeltaTableBuilder,character-method}
\alias{dlt_comment,InternalDeltaTableBuilder,character-method}
\title{Add comment describing the build table.}
\usage{
\S4method{dlt_comment}{InternalDeltaTableBuilder,character}(dtb, comment)
}
\arguments{
\item{dtb}{InternalDeltaTableBuilder}

\item{comment}{character, path}
}
\value{
this InternalDeltaTableBuilder
}
\description{
Add comment describing the build table.
}
\section{Functions}{
\itemize{
\item \code{dlt_comment,InternalDeltaTableBuilder,character-method}: Add comment
}}

\note{
dlt_comment, since 1.0.0
}
\keyword{internal}
