% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_comment,DeltaTableBuilder,character-method}
\alias{dlt_comment,DeltaTableBuilder,character-method}
\title{Add comment describing the build table.}
\usage{
\S4method{dlt_comment}{DeltaTableBuilder,character}(dtb, comment)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{comment}{character, path}
}
\value{
DeltaTableBuilder
}
\description{
Add comment describing the build table.
}
\section{Functions}{
\itemize{
\item \code{dlt_comment,DeltaTableBuilder,character-method}: Add comment
}}

\note{
dlt_comment, since 1.0.0
}
