% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_property,DeltaTableBuilder,character,character-method}
\alias{dlt_property,DeltaTableBuilder,character,character-method}
\title{Set property for the build table}
\usage{
\S4method{dlt_property}{DeltaTableBuilder,character,character}(dtb, key, value)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{key}{character}

\item{value}{character}
}
\description{
Set property for the build table
}
\section{Functions}{
\itemize{
\item \code{dlt_property,DeltaTableBuilder,character,character-method}: Set property
}}

\note{
dlt_property, since 1.0.0
}
