% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_location,DeltaTableBuilder,character-method}
\alias{dlt_location,DeltaTableBuilder,character-method}
\title{Specify data storage location for the build table.}
\usage{
\S4method{dlt_location}{DeltaTableBuilder,character}(dtb, location)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{location}{character, path}
}
\value{
DeltaTableBuilder
}
\description{
Specify data storage location for the build table.
}
\section{Functions}{
\itemize{
\item \code{dlt_location,DeltaTableBuilder,character-method}: Specify data storage location
}}

\note{
dlt_location, since 1.0.0
}
