% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_partitioned_by,DeltaTableBuilder-method}
\alias{dlt_partitioned_by,DeltaTableBuilder-method}
\title{Specify partitioning columns}
\usage{
\S4method{dlt_partitioned_by}{DeltaTableBuilder}(dtb, ...)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{...}{character columns}
}
\description{
Specify partitioning columns
}
\section{Functions}{
\itemize{
\item \code{dlt_partitioned_by,DeltaTableBuilder-method}: Specify partitioning
}}

\note{
dlt_partitioned_by, since 1.0.0
}
