% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build_internal.R, R/build.R
\name{dlt_add_columns}
\alias{dlt_add_columns}
\alias{dlt_add_columns,InternalDeltaTableBuilder,structType-method}
\alias{dlt_add_columns,InternalDeltaTableBuilder,character-method}
\alias{dlt_add_columns,DeltaTableBuilder,ANY-method}
\title{Add column to the build table}
\usage{
dlt_add_columns(dtb, schema)

\S4method{dlt_add_columns}{InternalDeltaTableBuilder,structType}(dtb, schema)

\S4method{dlt_add_columns}{InternalDeltaTableBuilder,character}(dtb, schema)

\S4method{dlt_add_columns}{DeltaTableBuilder,ANY}(dtb, schema)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{schema}{character (DDL string) or structType}
}
\value{
this InternalDeltaTableBuilder

DeltaTableBuilder
}
\description{
Add column to the build table

Add column to the build table
}
\note{
dlt_add_columns, since 1.0.0

dlt_add_columns, since 1.0.0
}
