% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_alias}
\alias{dlt_alias}
\alias{dlt_alias,DeltaTable,character-method}
\title{Set alias for this DeltaTable}
\usage{
dlt_alias(dt, name)

\S4method{dlt_alias}{DeltaTable,character}(dt, name)
}
\arguments{
\item{dt}{DeltaTable}

\item{name}{character}
}
\value{
DeltaTable
}
\description{
Set alias for this DeltaTable
}
\note{
dlt_alias, since 1.0.0
}
\seealso{
\linkS4class{DeltaTable}
}
