% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R
\name{dlt_comment}
\alias{dlt_comment}
\alias{dlt_comment,DeltaTableBuilder,character-method}
\title{Add comment describing the build table.}
\usage{
dlt_comment(dtb, comment)

\S4method{dlt_comment}{DeltaTableBuilder,character}(dtb, comment)
}
\arguments{
\item{dtb}{Delta table builder}

\item{comment}{character, path}
}
\value{
this DeltaTableBuilder
}
\description{
Add comment describing the build table.
}
\note{
dlt_comment, since 1.0.0
}
