% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_convert_to_delta}
\alias{dlt_convert_to_delta}
\alias{dlt_convert_to_delta,character,missing-method}
\alias{dlt_convert_to_delta,character,character-method}
\title{Create a DeltaTable from the given parquet table.}
\usage{
dlt_convert_to_delta(identifier, partition_schema)

\S4method{dlt_convert_to_delta}{character,missing}(identifier)

\S4method{dlt_convert_to_delta}{character,character}(identifier, partition_schema)
}
\arguments{
\item{identifier}{character}

\item{partition_schema}{optional, character}
}
\value{
DeltaTable
}
\description{
Create a DeltaTable from the given parquet table.
}
\note{
dlt_convert_to_delta, since 1.0.0
}
\examples{
\dontrun{
library(SparkR)
createDataFrame(iris) \%>\%
  write.parquet("/tmp/iris-non-delta")

dlt_convert_to_delta("parquet.`/tmp/iris-non-delta`") \%>\%
  dlt_to_df() \%>\%
  showDF()
}

}
\seealso{
\linkS4class{DeltaTable}
}
