% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_create_if_not_exists}
\alias{dlt_create_if_not_exists}
\title{Build DeltaTable if doesn't exits}
\usage{
dlt_create_if_not_exists()
}
\value{
DeltaTableBuilder
}
\description{
Build DeltaTable if doesn't exits
}
\note{
dlt_create_if_not_exists, since 1.0.0
}
