% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_delete}
\alias{dlt_delete}
\alias{dlt_delete,DeltaTable,missing-method}
\alias{dlt_delete,DeltaTable,Column-method}
\alias{dlt_delete,DeltaTable,character-method}
\title{Delete data from this DataTable}
\usage{
dlt_delete(dt, condition)

\S4method{dlt_delete}{DeltaTable,missing}(dt)

\S4method{dlt_delete}{DeltaTable,Column}(dt, condition)

\S4method{dlt_delete}{DeltaTable,character}(dt, condition)
}
\arguments{
\item{dt}{DeltaTable}

\item{condition}{optional, character of Column
If character, interpreted as SQL expression}
}
\description{
Removes data that matches optional condition
}
\note{
dlt_delete, since 1.0.0
}
