% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_detail}
\alias{dlt_detail}
\alias{dlt_detail,DeltaTable-method}
\title{Get the details of a Delta table such as the format, name, and size.}
\usage{
dlt_detail(dt)

\S4method{dlt_detail}{DeltaTable}(dt)
}
\arguments{
\item{dt}{DeltaTable}
}
\value{
SparkDataFrame containing information about the table
}
\description{
Get the details of a Delta table such as the format, name, and size.
}
\note{
dlt_detail, since 2.1.0
}
\seealso{
\linkS4class{DeltaTable}
}
