% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{dlt_execute,DeltaMergeBuilder-method}
\alias{dlt_execute,DeltaMergeBuilder-method}
\title{Execute merge operation on this builder}
\usage{
\S4method{dlt_execute}{DeltaMergeBuilder}(bldr)
}
\arguments{
\item{bldr}{DeltaMergeBuilder}
}
\value{
target DeltaTable, invisibly
}
\description{
Execute merge operation on this builder
}
\section{Functions}{
\itemize{
\item \code{dlt_execute(DeltaMergeBuilder)}: Execute merge operation

}}
\note{
dlt_execute, since 1.0.0
}
\seealso{
\linkS4class{DeltaMergeBuilder}
}
