% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/optimize.R
\name{dlt_execute_compaction}
\alias{dlt_execute_compaction}
\alias{dlt_execute_compaction,DeltaOptimizeBuilder-method}
\title{Compact the small files in selected partitions.}
\usage{
dlt_execute_compaction(dob)

\S4method{dlt_execute_compaction}{DeltaOptimizeBuilder}(dob)
}
\arguments{
\item{dob}{DeltaOptimizeBuilder}
}
\value{
SparkDataFrame containing the OPTIMIZE execution metrics
}
\description{
Compact the small files in selected partitions.
}
\note{
dlt_execute_compaction, since 2.0.0
}
\seealso{
\linkS4class{DeltaOptimizeBuilder}
}
