% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/optimize.R
\name{dlt_execute_z_order_by}
\alias{dlt_execute_z_order_by}
\alias{dlt_execute_z_order_by,DeltaOptimizeBuilder-method}
\title{Z-Order the data in selected partitions using the given columns.}
\usage{
dlt_execute_z_order_by(dob, ...)

\S4method{dlt_execute_z_order_by}{DeltaOptimizeBuilder}(dob, ...)
}
\arguments{
\item{dob}{DeltaOptimizeBuilder}

\item{...}{character columns}
}
\value{
SparkDataFrame containing the OPTIMIZE execution metrics
}
\description{
Z-Order the data in selected partitions using the given columns.
}
\note{
dlt_execute_z_order_by, since 2.0.0
}
\seealso{
\linkS4class{DeltaOptimizeBuilder}
}
