% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dlt_for_name}
\alias{dlt_for_name}
\title{Load DeltaTable from metastore}
\usage{
dlt_for_name(name)
}
\arguments{
\item{name}{character}
}
\value{
DeltaTable
}
\description{
Create DeltaTable for a given name
}
\note{
dlt_for_name, since 1.0.0
}
