% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{dlt_for_path}
\alias{dlt_for_path}
\title{Load DeltaTable from a path}
\usage{
dlt_for_path(path)
}
\arguments{
\item{path}{character}
}
\value{
DeltaTable
}
\description{
Create DeltaTable using given path
}
\note{
dlt_for_path, since 1.0.0
}
\seealso{
\linkS4class{DeltaTable}
}
