% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dlt_generate_manifest}
\alias{dlt_generate_manifest}
\alias{dlt_generate_manifest,DeltaTable,character-method}
\title{Generate manifest files for this DeltaTable}
\usage{
dlt_generate_manifest(dt, mode)

\S4method{dlt_generate_manifest}{DeltaTable,character}(dt, mode = c("symlink_format_manifest"))
}
\arguments{
\item{dt}{DeltaTable}

\item{mode}{character mode for the type of manifest file to be generated}
}
\description{
Generate manifest files for this DeltaTable
}
\note{
dlt_generate_manifest, since 1.0.0
}
