% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_history}
\alias{dlt_history}
\alias{dlt_history,DeltaTable,missing-method}
\alias{dlt_history,DeltaTable,numeric-method}
\title{Get commit history for this DeltaTable}
\usage{
dlt_history(dt, limit)

\S4method{dlt_history}{DeltaTable,missing}(dt)

\S4method{dlt_history}{DeltaTable,numeric}(dt, limit)
}
\arguments{
\item{dt}{DeltaTable}

\item{limit}{optional, numeric}
}
\value{
SparkDataFrame
}
\description{
Get commit history for this DeltaTable
}
\note{
dlt_history, since 1.0.0
}
