% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{dlt_initialize_delta_table_builder}
\alias{dlt_initialize_delta_table_builder}
\alias{dlt_create}
\alias{dlt_create_if_not_exists}
\alias{dlt_replace}
\alias{dlt_create_or_replace}
\title{Initalize new DeltaTableBuilder}
\usage{
dlt_create()

dlt_create_if_not_exists()

dlt_replace()

dlt_create_or_replace()
}
\value{
DeltaTableBuilder
}
\description{
Initalize new DeltaTableBuilder
}
\section{Functions}{
\itemize{
\item \code{dlt_create}: On execute, creates a new DeltaTable if it dosen't exist,
otherwise throws an exception.

\item \code{dlt_create_if_not_exists}: On execute, creates a new DeltaTable if it dosen't exist,
otherwise does nothing.

\item \code{dlt_replace}: On execute, replaces DeltaTable if it exists,
otherwise throws an exception.

\item \code{dlt_create_or_replace}: On execute, replaces DeltaTable if it exists,
otherwise creates a new DeltaTable
}}

\note{
dlt_create, since 1.0.0

dlt_create_if_not_exists, since 1.0.0

dlt_replace, since 1.0.0

dlt_create_or_replace, since 1.0.0
}
\seealso{
\linkS4class{DeltaTableBuilder}
}
