% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_is_delta_table}
\alias{dlt_is_delta_table}
\alias{dlt_is_delta_table,character-method}
\title{Check if path corresponds to a DeltaTable}
\usage{
dlt_is_delta_table(identifier)

\S4method{dlt_is_delta_table}{character}(identifier)
}
\arguments{
\item{identifier}{character path}
}
\value{
logical
}
\description{
Check if path corresponds to a DeltaTable
}
\note{
dlt_is_delta_table, since 1.0.0
}
