% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\name{dlt_location,InternalDeltaTableBuilder,character-method}
\alias{dlt_location,InternalDeltaTableBuilder,character-method}
\title{Specify data storage location for the build table.}
\usage{
\S4method{dlt_location}{InternalDeltaTableBuilder,character}(dtb, location)
}
\arguments{
\item{dtb}{InternalDeltaTableBuilder}

\item{location}{character, path}
}
\value{
this InternalDeltaTableBuilder
}
\description{
Specify data storage location for the build table.
}
\section{Functions}{
\itemize{
\item \code{dlt_location(dtb = InternalDeltaTableBuilder, location = character)}: Specify data storage location

}}
\note{
dlt_location, since 1.0.0
}
\keyword{internal}
