% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R
\name{dlt_location}
\alias{dlt_location}
\alias{dlt_location,DeltaTableBuilder,character-method}
\title{Specify data storage location for the build table.}
\usage{
dlt_location(dtb, location)

\S4method{dlt_location}{DeltaTableBuilder,character}(dtb, location)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{location}{character, path}
}
\value{
DeltaTableBuilder
}
\description{
Specify data storage location for the build table.
}
\note{
dlt_location, since 1.0.0
}
\seealso{
\linkS4class{DeltaTableBuilder}
}
