% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_merge}
\alias{dlt_merge}
\alias{dlt_merge,DeltaTable,SparkDataFrame,Column-method}
\alias{dlt_merge,DeltaTable,SparkDataFrame,character-method}
\title{Merge data from sourc with this DataTable, based on condition}
\usage{
dlt_merge(dt, source, condition)

\S4method{dlt_merge}{DeltaTable,SparkDataFrame,Column}(dt, source, condition)

\S4method{dlt_merge}{DeltaTable,SparkDataFrame,character}(dt, source, condition)
}
\arguments{
\item{dt}{DeltaTable}

\item{source}{SparkDataFrame}

\item{condition}{character or Column}
}
\description{
Merge data from sourc with this DataTable, based on condition
}
\note{
dlt_merge, since 1.0.0
}
