% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_optimize}
\alias{dlt_optimize}
\alias{dlt_optimize,DeltaTable-method}
\title{Optimize the data layout of the table.}
\usage{
dlt_optimize(dt)

\S4method{dlt_optimize}{DeltaTable}(dt)
}
\arguments{
\item{dt}{DeltaTable}
}
\value{
DeltaOptimizeBuilder
}
\description{
Optimize the data layout of the table.
}
\note{
dlt_optimize, since 2.0.0
}
\seealso{
\linkS4class{DeltaOptimizeBuilder}
}
