% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\name{dlt_partitioned_by,InternalDeltaTableBuilder-method}
\alias{dlt_partitioned_by,InternalDeltaTableBuilder-method}
\title{Specify partitioning columns}
\usage{
\S4method{dlt_partitioned_by}{InternalDeltaTableBuilder}(dtb, ...)
}
\arguments{
\item{dtb}{InternalDeltaTableBuilder}

\item{...}{character columns}
}
\description{
Specify partitioning columns
}
\section{Functions}{
\itemize{
\item \code{dlt_partitioned_by(InternalDeltaTableBuilder)}: Specify partitioning

}}
\note{
dlt_partitioned_by, since 1.0.0
}
\keyword{internal}
