% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R
\name{dlt_partitioned_by}
\alias{dlt_partitioned_by}
\alias{dlt_partitioned_by,DeltaTableBuilder-method}
\title{Specify partitioning columns}
\usage{
dlt_partitioned_by(dtb, ...)

\S4method{dlt_partitioned_by}{DeltaTableBuilder}(dtb, ...)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{...}{character columns}
}
\description{
Specify partitioning columns
}
\note{
dlt_partitioned_by, since 1.0.0
}
\seealso{
\linkS4class{DeltaTableBuilder}
}
