% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{dlt_read}
\alias{dlt_read}
\title{Create a SparkDataFrame from a Delta Table}
\usage{
dlt_read(path, ...)
}
\arguments{
\item{path}{path of file to read. A vector of multiple paths is allowed.}

\item{...}{additional data source specific named properties.}
}
\description{
Loads a DeltaTable, returning the result as a SparkDataFrame.
}
\note{
dlt_read since 1.0.0
}
\examples{
\dontrun{
dlt_read("/tmp/iris-delta")
}
}
