% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite_stream.R
\name{dlt_read_stream}
\alias{dlt_read_stream}
\title{Create a streaming SparkDataFrame from a Delta directory}
\usage{
dlt_read_stream(path, ...)
}
\arguments{
\item{path}{path of file to read.}

\item{...}{additional data source specific named properties.}
}
\description{
Loads a DeltaTable, returning the result as a streaming SparkDataFrame.
}
\note{
dlt_read_stream since 1.0.0
}
\examples{
\dontrun{
dlt_read_stream("/tmp/iris-delta")
}
}
