% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/readwrite.R
\name{dlt_write}
\alias{dlt_write}
\alias{write.delta}
\alias{dlt_readwrite}
\alias{dlt_read}
\alias{read.delta}
\alias{dlt_write,SparkDataFrame,character-method}
\alias{write.delta,SparkDataFrame,character-method}
\title{Write and read SparkDataFrame in delta format}
\usage{
dlt_write(df, path, ...)

write.delta(df, path, ...)

dlt_read(path, ...)

read.delta(path, ...)

\S4method{dlt_write}{SparkDataFrame,character}(df, path, ...)

\S4method{write.delta}{SparkDataFrame,character}(df, path, ...)
}
\arguments{
\item{df}{SparkDataFrame}

\item{path}{character path to write the data.}

\item{...}{additional arguments passed to writer}
}
\description{
Write and read SparkDataFrame in delta format
}
\section{Functions}{
\itemize{
\item \code{dlt_write}: Generic

\item \code{write.delta}: Generic

\item \code{dlt_read}: Load data stored as delta into \code{SparkDataFrame}

\item \code{read.delta}: Load data stored as delta into \code{SparkDataFrame} (alias of \code{dlt_read})

\item \code{dlt_write,SparkDataFrame,character-method}: Write SparkDataFrame` in delta format

\item \code{write.delta,SparkDataFrame,character-method}: Write \code{SparkDataFrame} in delta format (alias of \code{dlt_write})
}}

\note{
dlt_read since 1.0.0

read.delta since 1.0.0

dlt_write since 1.0.0

write.delta since 1.0.0
}
\examples{
\dontrun{
set.seed(323)
path <- tempfile()

df <- data.frame(
  id = 1:12,
  key = rep(c("a", "b", "c"), each = 4),
  value = rnorm(12)
) \%>\%
  createDataFrame()

dlt_write(df, path)

dlt_read(path) \%>\%
  schema()

write.delta(df, path, mode = "overwrite")

read.delta(path) \%>\%
  schema()
}

}
