% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/readwrite_stream.R
\name{dlt_write_stream}
\alias{dlt_write_stream}
\alias{dlt_readwrite_stream}
\alias{dlt_read_stream}
\alias{dlt_write_stream,SparkDataFrame,character-method}
\title{Write and read streaming SparkDataFrame in delta format}
\usage{
dlt_write_stream(df, path, ...)

dlt_read_stream(path, ...)

\S4method{dlt_write_stream}{SparkDataFrame,character}(df, path, ...)
}
\arguments{
\item{df}{streaming SparkDataFrame}

\item{path}{character path to write the data.}

\item{...}{additional arguments passed to writer}
}
\description{
Write and read streaming SparkDataFrame in delta format
}
\section{Functions}{
\itemize{
\item \code{dlt_write_stream()}: Generic

\item \code{dlt_read_stream()}: Loads data stored in delta format, returning the result as a streaming SparkDataFrame

\item \code{dlt_write_stream(df = SparkDataFrame, path = character)}: Writes streaming SparkDataFrame using delta format

}}
\note{
dlt_read_stream since 1.0.0

dlt_write_stream since 1.0.0
}
\examples{
\dontrun{
set.seed(323)
input_path <- tempfile()
output_path <- tempfile()

# Dummy input
data.frame(
  id = 1:12,
  key = rep(c("a", "b", "c"), each = 4),
  value = rnorm(12)
) \%>\%
  createDataFrame() \%>\%
  dlt_write(input_path)

# Read data as stream and write back to output location
query <- dlt_read_stream(path = input_path) \%>\%
  dlt_write_stream(
    path = output_path, queryName = "test", trigger.once = TRUE,
    checkpointLocation = file.path(output_path, "_checkpoints", "test")
  )

awaitTermination(query)
}

}
