% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_restore_to_timestamp}
\alias{dlt_restore_to_timestamp}
\alias{dlt_restore_to_timestamp,DeltaTable,character-method}
\title{Restore the DeltaTable to an older version of the table specified by a timestamp.}
\usage{
dlt_restore_to_timestamp(dt, timestamp)

\S4method{dlt_restore_to_timestamp}{DeltaTable,character}(dt, timestamp)
}
\arguments{
\item{dt}{DeltaTable}

\item{timestamp}{character formatted as yyyy-MM-dd or yyyy-MM-dd HH:mm:ss}
}
\value{
SparkDataFrame
}
\description{
Timestamp can be of the format yyyy-MM-dd or yyyy-MM-dd HH:mm:ss
}
\note{
dlt_restore_to_timestamp, since 1.2.0
}
\seealso{
\linkS4class{DeltaTable}
}
