% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_restore_to_version}
\alias{dlt_restore_to_version}
\alias{dlt_restore_to_version,DeltaTable,numeric-method}
\title{Restore the DeltaTable to an older version of the table specified by version number.}
\usage{
dlt_restore_to_version(dt, version)

\S4method{dlt_restore_to_version}{DeltaTable,numeric}(dt, version)
}
\arguments{
\item{dt}{DeltaTable}

\item{version}{int}
}
\value{
SparkDataFrame
}
\description{
Restore the DeltaTable to an older version of the table specified by version number.
}
\note{
dlt_restore_to_version, since 1.2.0
}
\seealso{
\linkS4class{DeltaTable}
}
