% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/show.R
\name{dlt_show}
\alias{dlt_show}
\alias{dlt_show,DeltaTable-method}
\title{Print rows of the corresponding SparkDataFrame}
\usage{
dlt_show(dt, ...)

\S4method{dlt_show}{DeltaTable}(dt, ...)
}
\arguments{
\item{dt}{DeltaTable}

\item{...}{Additional arguments passed to \code{SparkR::showDF}}
}
\value{
this DeltaTable, invisibly
}
\description{
Convenience method equivalent to
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{dt \%>\%
 dlt_to_df() \%>\%
 showDF()
}\if{html}{\out{</div>}}
}
\note{
dlt_show, since 1.0.0
}
