% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\name{dlt_table_name,InternalDeltaTableBuilder,character-method}
\alias{dlt_table_name,InternalDeltaTableBuilder,character-method}
\title{Specify name of the build table.}
\usage{
\S4method{dlt_table_name}{InternalDeltaTableBuilder,character}(dtb, identifier)
}
\arguments{
\item{dtb}{InternalDeltaTableBuilder}

\item{identifier}{character}
}
\value{
this InternalDeltaTableBuilder
}
\description{
Specify name of the build table.
}
\section{Functions}{
\itemize{
\item \code{dlt_table_name,InternalDeltaTableBuilder,character-method}: Specify name of the table
}}

\note{
dlt_table_name, since 1.0.0
}
\keyword{internal}
