% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_to_df}
\alias{dlt_to_df}
\alias{dlt_to_df,DeltaTable-method}
\title{Get SparkDataFrame corresponding to DeltaTable}
\usage{
dlt_to_df(dt)

\S4method{dlt_to_df}{DeltaTable}(dt)
}
\arguments{
\item{dt}{DeltaTable}
}
\value{
SparkDataFrame
}
\description{
Get SparkDataFrame corresponding to DeltaTable
}
\note{
dlt_to_df, since 1.0.0
}
