% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_update}
\alias{dlt_update}
\alias{dlt_update,DeltaTable,characterOrList,missing-method}
\alias{dlt_update,DeltaTable,characterOrList,Column-method}
\alias{dlt_update,DeltaTable,characterOrList,character-method}
\title{Update data in this DataTable}
\usage{
dlt_update(dt, set, condition)

\S4method{dlt_update}{DeltaTable,characterOrList,missing}(dt, set)

\S4method{dlt_update}{DeltaTable,characterOrList,Column}(dt, set, condition)

\S4method{dlt_update}{DeltaTable,characterOrList,character}(dt, set, condition)
}
\arguments{
\item{dt}{DeltaTable}

\item{set}{named character vector or list of Columns}

\item{condition}{optional, character or Column}
}
\description{
Update data in this DataTable
}
\note{
dlt_update, since 1.0.0
}
