% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_upgrade_table_protocol}
\alias{dlt_upgrade_table_protocol}
\alias{dlt_upgrade_table_protocol,DeltaTable,numeric,numeric-method}
\title{Upgrade protocol version of this DeltaTable}
\usage{
dlt_upgrade_table_protocol(dt, reader_version, writer_version)

\S4method{dlt_upgrade_table_protocol}{DeltaTable,numeric,numeric}(dt, reader_version, writer_version)
}
\arguments{
\item{dt}{DeltaTable}

\item{reader_version}{numeric}

\item{writer_version}{numeric}
}
\value{
NULL, invisibly
}
\description{
Upgrade protocol version of this DeltaTable
}
\note{
dlt_upgrade_table_protocol, since 1.0.0
}
