% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/table.R
\name{dlt_vacuum}
\alias{dlt_vacuum}
\alias{dlt_vacuum,DeltaTable,missing-method}
\alias{dlt_vacuum,DeltaTable,numeric-method}
\title{Vacuum this DeltaTable}
\usage{
dlt_vacuum(dt, retention_hours)

\S4method{dlt_vacuum}{DeltaTable,missing}(dt)

\S4method{dlt_vacuum}{DeltaTable,numeric}(dt, retention_hours)
}
\arguments{
\item{dt}{DataTable}

\item{retention_hours}{numeric}
}
\value{
NULL, invisibly
}
\description{
Vacuum this DeltaTable
}
\note{
dlt_vacuum, since 1.0.0
}
