% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/merge.R
\name{dlt_when_matched_update}
\alias{dlt_when_matched_update}
\alias{dlt_when_matched_update,DeltaMergeBuilder,characterOrList,missing-method}
\alias{dlt_when_matched_update,DeltaMergeBuilder,characterOrList,Column-method}
\alias{dlt_when_matched_update,DeltaMergeBuilder,characterOrList,character-method}
\title{Update rows for table using set rules}
\usage{
dlt_when_matched_update(dmb, set, condition)

\S4method{dlt_when_matched_update}{DeltaMergeBuilder,characterOrList,missing}(dmb, set)

\S4method{dlt_when_matched_update}{DeltaMergeBuilder,characterOrList,Column}(dmb, set, condition)

\S4method{dlt_when_matched_update}{DeltaMergeBuilder,characterOrList,character}(dmb, set, condition)
}
\arguments{
\item{dmb}{DeltaMergeBuilder}

\item{set}{character or list}

\item{condition}{optional, character or Column}
}
\value{
DeltaMergeBuilder
}
\description{
Update rows for table using set rules
}
\note{
dlt_when_matched_update, since 1.0.0
}
\seealso{
\linkS4class{DeltaMergeBuilder}
}
