% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/merge.R
\name{dlt_when_not_matched_insert}
\alias{dlt_when_not_matched_insert}
\alias{dlt_when_not_matched_insert,DeltaMergeBuilder,characterOrList,missing-method}
\alias{dlt_when_not_matched_insert,DeltaMergeBuilder,characterOrList,Column-method}
\alias{dlt_when_not_matched_insert,DeltaMergeBuilder,characterOrList,character-method}
\title{Insert a new row based the set of rules}
\usage{
dlt_when_not_matched_insert(dmb, set, condition)

\S4method{dlt_when_not_matched_insert}{DeltaMergeBuilder,characterOrList,missing}(dmb, set)

\S4method{dlt_when_not_matched_insert}{DeltaMergeBuilder,characterOrList,Column}(dmb, set, condition)

\S4method{dlt_when_not_matched_insert}{DeltaMergeBuilder,characterOrList,character}(dmb, set, condition)
}
\arguments{
\item{dmb}{DeltaMergeBuilder}

\item{set}{character or list}

\item{condition}{optional, character or Column}
}
\value{
DeltaMergeBuilder
}
\description{
Insert a new row based the set of rules
}
\note{
dlt_when_not_matched_insert, since 1.0.0
}
