% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/merge.R
\name{dlt_when_not_matched_insert_all}
\alias{dlt_when_not_matched_insert_all}
\alias{dlt_when_not_matched_insert_all,DeltaMergeBuilder,missing-method}
\alias{dlt_when_not_matched_insert_all,DeltaMergeBuilder,Column-method}
\alias{dlt_when_not_matched_insert_all,DeltaMergeBuilder,character-method}
\title{Insert a new row if not source not matched}
\usage{
dlt_when_not_matched_insert_all(dmb, condition)

\S4method{dlt_when_not_matched_insert_all}{DeltaMergeBuilder,missing}(dmb)

\S4method{dlt_when_not_matched_insert_all}{DeltaMergeBuilder,Column}(dmb, condition)

\S4method{dlt_when_not_matched_insert_all}{DeltaMergeBuilder,character}(dmb, condition)
}
\arguments{
\item{dmb}{DeltaMergeBuilder}

\item{condition}{optional, character or Column}
}
\value{
DeltaMergeBuilder
}
\description{
Insert a new row if not source not matched
}
\note{
dlt_when_not_matched_insert_all, since 1.0.0
}
\seealso{
\linkS4class{DeltaMergeBuilder}
}
