% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/readwrite.R
\name{dlt_write}
\alias{dlt_write}
\alias{dlt_write,SparkDataFrame,character-method}
\title{Write SparkDataFrame to Delta}
\usage{
dlt_write(df, path, ...)

\S4method{dlt_write}{SparkDataFrame,character}(df, path, ...)
}
\arguments{
\item{df}{SparkDataFrame}

\item{path}{character path to write the data.}

\item{...}{additional arguments passed to writer}
}
\description{
Writes SparkDataFrame to Delta.
}
\note{
dlt_write since 1.0.0
}
\examples{
\dontrun{
SparkR::createDataFrame(iris) \%>\%
  dlt_write("/tmp/iris-delta")
}
}
