% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/readwrite_stream.R
\name{dlt_write_stream}
\alias{dlt_write_stream}
\alias{dlt_write_stream,SparkDataFrame,character-method}
\title{Write streaming SparkDataFrame to Delta}
\usage{
dlt_write_stream(df, path, ...)

\S4method{dlt_write_stream}{SparkDataFrame,character}(df, path, ...)
}
\arguments{
\item{df}{streaming SparkDataFrame}

\item{path}{character path to write the data.}

\item{...}{additional arguments passed to writer}
}
\description{
Writes streaming SparkDataFrame to Delta.
}
\note{
dlt_write_stream since 1.0.0
}
