% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_internal.R
\docType{class}
\name{InternalDeltaTableBuilder-class}
\alias{InternalDeltaTableBuilder-class}
\title{S4 class that represents pending table build operation}
\description{
This class is used primarily to support \code{DeltaTableBuilder}.
Internally, it holds reference to a mutable Java \code{DeltaTableBuilder}
and all implemented methods mutate it in place.
While it is possible to use it directly, \code{DeltaTableBuilder}
is preferred and provides the same API.
}
\section{Slots}{

\describe{
\item{\code{jtb}}{A Java object reference to the backing InternalDeltaTableBuilder}
}}

\note{
InternalDeltaTableBuilder, since 1.0.0
}
\seealso{
DeltaTableBuilder
}
\concept{InternalDeltaTableBuilder functions}
\keyword{internal}
