% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readwrite.R
\name{read.delta}
\alias{read.delta}
\title{Create a SparkDataFrame from a Delta Table}
\usage{
read.delta(path, ...)
}
\arguments{
\item{path}{path of file to read.}

\item{...}{additional data source specific named properties.}
}
\description{
Loads a DeltaTable, returning the result as a SparkDataFrame.
}
\note{
read.delta since 1.0.0
}
\examples{
\dontrun{
read.delta("/tmp/iris-delta")
}
}
\seealso{
dlt_read
}
