% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/readwrite.R
\name{write.delta}
\alias{write.delta}
\alias{write.delta,SparkDataFrame,character-method}
\title{Write SparkDataFrame to Delta}
\usage{
write.delta(df, path, ...)

\S4method{write.delta}{SparkDataFrame,character}(df, path, ...)
}
\arguments{
\item{df}{SparkDataFrame}

\item{path}{character path to write the data.}

\item{...}{additional arguments passed to writer}
}
\description{
Writes SparkDataFrame to Delta.
}
\note{
write.delta since 1.0.0
}
\examples{
\dontrun{
SparkR::createDataFrame(iris) \%>\%
  write.delta("/tmp/iris-delta")
}
}
\seealso{
dlt_write
}
