% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\docType{class}
\name{DeltaOptimizeBuilder-class}
\alias{DeltaOptimizeBuilder-class}
\title{S4 class that represents pending optimize operation}
\description{
S4 class that represents pending optimize operation
}
\section{Slots}{

\describe{
\item{\code{job}}{A Java object reference to the backing DeltaOptimizeBuilder}
}}

\note{
DeltaOptimizeBuilder, since 2.0.0
}
\examples{
\dontrun{

}

}
\concept{DeltaOptimizeBuilder functions}
