% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\docType{class}
\name{DeltaTableBuilder-class}
\alias{DeltaTableBuilder-class}
\title{S4 class that represents pending table build operation}
\description{
S4 class that represents pending table build operation
}
\section{Slots}{

\describe{
\item{\code{jtb}}{A Java object reference to the backing DeltaTableBuilder}
}}

\note{
DeltaTableBuilder, since 1.0.0
}
\concept{DeltaTableBuilder functions}
