% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\docType{class}
\name{DeltaTableBuilder-class}
\alias{DeltaTableBuilder-class}
\title{S4 class that represents pending table build operation}
\description{
S4 class that represents pending table build operation
}
\section{Slots}{

\describe{
\item{\code{method}}{character name of the initializer method}

\item{\code{ops}}{list list of operations in form of list(.method = ..., args = list(...))}
}}

\note{
DeltaTableBuilder, since 1.0.0
}
\concept{DeltaTableBuilder functions}
