% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\docType{class}
\name{DeltaTableBuilder-class}
\alias{DeltaTableBuilder-class}
\title{S4 class that represents pending table build operation}
\description{
This class is used primarily to support \code{PureDeltaTableBuilder}.
Internally, it holds reference to a mutable Java \code{DeltaTableBuilder}
and all implemented methods mutate it in place.
While it is possible to use it directly, \code{PureDeltaTableBuilder}
is preferred and provides the same API.
}
\section{Slots}{

\describe{
\item{\code{jtb}}{A Java object reference to the backing DeltaTableBuilder}
}}

\note{
DeltaTableBuilder, since 1.0.0
}
\seealso{
PureDeltaTableBuilder
}
\concept{DeltaTableBuilder functions}
