% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R, R/build_pure.R
\name{dlt_add_columns}
\alias{dlt_add_columns}
\alias{dlt_add_columns,DeltaTableBuilder,structType-method}
\alias{dlt_add_columns,DeltaTableBuilder,character-method}
\alias{dlt_add_columns,PureDeltaTableBuilder,ANY-method}
\title{Add column to the build table}
\usage{
dlt_add_columns(dtb, schema)

\S4method{dlt_add_columns}{DeltaTableBuilder,structType}(dtb, schema)

\S4method{dlt_add_columns}{DeltaTableBuilder,character}(dtb, schema)

\S4method{dlt_add_columns}{PureDeltaTableBuilder,ANY}(dtb, schema)
}
\arguments{
\item{dtb}{PureDeltaTableBuilder}

\item{schema}{character (DDL string) or structType}
}
\value{
this DeltaTableBuilder

PureDeltaTableBuilder
}
\description{
Add column to the build table

Add column to the build table
}
\note{
dlt_add_columns, since 1.0.0

dlt_add_columns, since 1.0.0
}
