% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R, R/build_pure.R
\name{dlt_comment}
\alias{dlt_comment}
\alias{dlt_comment,DeltaTableBuilder,character-method}
\alias{dlt_comment,PureDeltaTableBuilder,character-method}
\title{Add comment describing the build table.}
\usage{
dlt_comment(dtb, comment)

\S4method{dlt_comment}{DeltaTableBuilder,character}(dtb, comment)

\S4method{dlt_comment}{PureDeltaTableBuilder,character}(dtb, comment)
}
\arguments{
\item{dtb}{PureDeltaTableBuilder}

\item{comment}{character, path}
}
\value{
this DeltaTableBuilder

PureDeltaTableBuilder
}
\description{
Add comment describing the build table.

Add comment describing the build table.
}
\note{
dlt_comment, since 1.0.0

dlt_comment, since 1.0.0
}
