% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build_internal.R, R/build.R,
%   R/merge.R
\name{dlt_execute}
\alias{dlt_execute}
\alias{dlt_execute,InternalDeltaTableBuilder-method}
\alias{dlt_execute,DeltaTableBuilder-method}
\alias{dlt_execute,DeltaMergeBuilder-method}
\title{Execute merge operation on this builder}
\usage{
dlt_execute(bldr)

\S4method{dlt_execute}{InternalDeltaTableBuilder}(bldr)

\S4method{dlt_execute}{DeltaTableBuilder}(bldr)

\S4method{dlt_execute}{DeltaMergeBuilder}(bldr)
}
\arguments{
\item{bldr}{DeltaMergeBuilder}
}
\value{
DeltaTable

DeltaTable

target DeltaTable, invisibly
}
\description{
Execute merge operation on this builder

Execute merge operation on this builder

Execute merge operation on this builder
}
\note{
dlt_execute, since 1.0.0

dlt_execute, since 1.0.0

dlt_execute, since 1.0.0
}
