% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dlt_execute}
\alias{dlt_execute}
\title{Execute pending operation on a builder}
\usage{
dlt_execute(bldr)
}
\arguments{
\item{bldr}{a builder object}
}
\description{
Execute pending operation on a builder
}
\keyword{internal}
