% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R, R/build_pure.R
\name{dlt_location}
\alias{dlt_location}
\alias{dlt_location,DeltaTableBuilder,character-method}
\alias{dlt_location,PureDeltaTableBuilder,character-method}
\title{Specify data storage location for the build table.}
\usage{
dlt_location(dtb, location)

\S4method{dlt_location}{DeltaTableBuilder,character}(dtb, location)

\S4method{dlt_location}{PureDeltaTableBuilder,character}(dtb, location)
}
\arguments{
\item{dtb}{PureDeltaTableBuilder}

\item{location}{character, path}
}
\value{
this DeltaTableBuilder

PureDeltaTableBuilder
}
\description{
Specify data storage location for the build table.

Specify data storage location for the build table.
}
\note{
dlt_location, since 1.0.0

dlt_location, since 1.0.0
}
