% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R, R/build_pure.R
\name{dlt_partitioned_by}
\alias{dlt_partitioned_by}
\alias{dlt_partitioned_by,DeltaTableBuilder-method}
\alias{dlt_partitioned_by,PureDeltaTableBuilder-method}
\title{Specify partitioning columns}
\usage{
dlt_partitioned_by(dtb, ...)

\S4method{dlt_partitioned_by}{DeltaTableBuilder}(dtb, ...)

\S4method{dlt_partitioned_by}{PureDeltaTableBuilder}(dtb, ...)
}
\arguments{
\item{dtb}{PureDeltaTableBuilder}

\item{...}{character columns}
}
\description{
Specify partitioning columns

Specify partitioning columns
}
\note{
dlt_partitioned_by, since 1.0.0

dlt_partitioned_by, since 1.0.0
}
