% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R
\name{dlt_property}
\alias{dlt_property}
\alias{dlt_property,DeltaTableBuilder,character,character-method}
\title{Set property for the build table}
\usage{
dlt_property(dtb, key, value)

\S4method{dlt_property}{DeltaTableBuilder,character,character}(dtb, key, value)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{key}{character}

\item{value}{character}
}
\description{
Set property for the build table
}
\note{
dlt_property, since 1.0.0
}
\seealso{
\linkS4class{DeltaTableBuilder}
}
