% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/build.R
\name{dlt_table_name}
\alias{dlt_table_name}
\alias{dlt_table_name,DeltaTableBuilder,character-method}
\title{Specify name of the build table.}
\usage{
dlt_table_name(dtb, identifier)

\S4method{dlt_table_name}{DeltaTableBuilder,character}(dtb, identifier)
}
\arguments{
\item{dtb}{DeltaTableBuilder}

\item{identifier}{character}
}
\value{
DeltaTableBuilder
}
\description{
Specify name of the build table.
}
\note{
dlt_table_name, since 1.0.0
}
\seealso{
\linkS4class{DeltaTableBuilder}
}
