% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/optimize.R
\name{dlt_where}
\alias{dlt_where}
\alias{dlt_where,DeltaOptimizeBuilder,character-method}
\title{Optimize the data layout of the table.}
\usage{
dlt_where(dob, partition_filter)

\S4method{dlt_where}{DeltaOptimizeBuilder}(dob, partition_filter)
}
\arguments{
\item{dob}{DeltaOptimizeBuilder}
}
\value{
DeltaOptimizeBuilder
}
\description{
Optimize the data layout of the table.
}
\note{
dlt_where, since 2.0.0
}
\seealso{
\linkS4class{DeltaOptimizeBuilder}
}
